+++ DDLs/DDL_DanmuRecord.sql
@@ -2,7 +2,7 @@ CREATE TABLE if not exists DanmuRecord
 (
     danmu_id serial primary key,
     bv       VARCHAR(255) REFERENCES VideoRecord (bv), -- video's BV for the danmu
-    mid      INT REFERENCES UserRecord (mid),          -- sender's mid of the danmu
+    mid      bigint REFERENCES UserRecord (mid),          -- sender's mid of the danmu
     time     float,                                    -- display time of the danmu in seconds since video starts
     content  TEXT,                                     -- content of the danmu
     postTime TIMESTAMP,                                -- post time of the danmu
